+++ jest.config.js
@@ -1,16 +1,24 @@
 module.exports = {
-  collectCoverageFrom: ['<rootDir>/src/**/*.{js,jsx}', '!<rootDir>/src/*.js'],
-  testMatch: ['<rootDir>/tests/**/*.{js,jsx}'],
-  transform: {
-    '^.+\\.(js|jsx)$': '<rootDir>/node_modules/babel-jest',
-  },
-  transformIgnorePatterns: ['[/\\\\]node_modules[/\\\\].+\\.(js|jsx)$'],
+  collectCoverageFrom: [
+    '<rootDir>/src/**/*.{ts,tsx}',
+    '!<rootDir>/src/*.{ts,tsx}',
+    '!<rootDir>/src/**/*.d.ts',
+  ],
   moduleNameMapper: {
     '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$':
       '<rootDir>/jest.mock-file.js',
     '\\.css$': 'identity-obj-proxy',
   },
-  moduleFileExtensions: ['js', 'json', 'jsx'],
-  setupFilesAfterEnv: ['jest-enzyme'],
-  testEnvironment: 'enzyme',
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json'],
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  testMatch: ['<rootDir>/tests/**/*.{ts,tsx}'],
+  transform: {
+    '^.+\\.(js|jsx)$': '<rootDir>/node_modules/babel-jest',
+  },
+  globals: {
+    'ts-jest': {
+      diagnostics: false,
+    },
+  },
 };