+++ src/components/SvgIcon/styles.css.d.ts
@@ -1,0 +1,4 @@
+export const Icon: string;
+export const icon: string;
+export const ProblemMessage: string;
+export const problemMessage: string;