+++ tests/components/SpriteIcon/index.test.js
@@ -1,9 +1,0 @@
-import SpriteIcon from '../../../src/components/SpriteIcon';
-
-describe('SpriteIcon component', () => {
-  it('should render something', () => {
-    const component = shallow(<SpriteIcon name="test" />);
-
-    expect(component).not.toBeEmptyRender();
-  });
-});