+++ webpack.config.js
@@ -6,10 +6,12 @@ const CleanWebpackPlugin = require('clean-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const WebappWebpackPlugin = require('webapp-webpack-plugin');
 const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
-const SpritesmithPlugin = require('webpack-spritesmith');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
+const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
+const { WatchIgnorePlugin } = require('webpack');
 
 const ENV = process.env.NODE_ENV || 'production';
+const ANALYZE_BUILD = process.env.ANALYZE_BUILD === '1';
 const APP_TITLE = process.env.APP_TITLE || 'React boilerplate';
 const APP_DESCRIPTION =
   process.env.APP_DESCRIPTION ||
@@ -26,49 +28,61 @@ const stats = {
   usedExports: false,
   modules: false,
 };
+const babelLoader = {
+  loader: 'babel-loader',
+  options: isDev
+    ? {}
+    : { cacheDirectory: true, cacheCompression: true, compact: true },
+};
 
 module.exports = {
   mode: ENV,
-  entry: './src/index.js',
+  entry: './src/index.tsx',
   output: {
     filename: '[name].[chunkhash].js',
     path: path.resolve(__dirname, 'dist'),
     publicPath: '/',
   },
   resolve: {
-    extensions: ['.js', '.json', '.jsx'],
+    extensions: ['.ts', '.tsx', '.js', '.jsx', '.json'],
   },
   module: {
     rules: [
       {
-        test: /\.(js|jsx)$/,
+        test: /\.(ts|tsx|js|jsx)$/,
         enforce: 'pre',
-        loader: 'eslint-loader',
+        loader: 'tslint-loader',
         include: path.resolve(__dirname, 'src'),
       },
       {
         oneOf: [
           {
-            test: /\.(js|jsx)$/,
-            loader: 'babel-loader',
+            test: /\.(ts|tsx)$/,
             include: path.resolve(__dirname, 'src'),
-            options: isDev
-              ? {}
-              : {
-                  cacheDirectory: true,
-                  cacheCompression: true,
-                  compact: true,
-                },
+            use: [
+              babelLoader,
+              {
+                loader: 'ts-loader',
+              },
+            ],
+          },
+          {
+            test: /\.(js|jsx)$/,
+            include: path.resolve(__dirname, 'src'),
+            use: [babelLoader],
           },
           {
             test: /\.css$/,
             use: [
               isDev ? 'style-loader' : MiniCssExtractPlugin.loader,
               {
-                loader: 'css-loader',
+                loader: 'typings-for-css-modules-loader',
                 options: {
                   modules: true,
                   importLoaders: 1,
+                  namedExport: true,
+                  camelCase: true,
+                  exportOnlyLocals: true,
                 },
               },
               'postcss-loader',
@@ -87,6 +101,7 @@ module.exports = {
     ],
   },
   plugins: [
+    new WatchIgnorePlugin([/css\.d\.ts$/]),
     new CleanWebpackPlugin(),
     new MiniCssExtractPlugin({
       filename: isDev ? '[name].css' : '[name].[contenthash].css',
@@ -113,29 +128,15 @@ module.exports = {
       },
     }),
     new OptimizeCSSAssetsPlugin(),
-    new SpritesmithPlugin({
-      src: {
-        cwd: path.resolve(__dirname, 'src/assets/sprites'),
-        glob: '*.png',
-      },
-      target: {
-        image: path.resolve(__dirname, 'public/sprite.png'),
-        css: path.resolve(__dirname, 'src/styles/sprite.css'),
-      },
-      apiOptions: {
-        cssImageRef: '/sprite.png',
-      },
-    }),
     new CopyWebpackPlugin([
       {
         from: path.resolve(__dirname, 'public/robots.txt'),
         to: path.resolve(__dirname, 'dist'),
       },
-      {
-        from: path.resolve(__dirname, 'public/sprite.png'),
-        to: path.resolve(__dirname, 'dist'),
-      },
     ]),
+    new BundleAnalyzerPlugin({
+      analyzerMode: ANALYZE_BUILD ? 'static' : 'disabled',
+    }),
   ],
   devtool: isDev ? 'source-map' : false,
   devServer: {